+++ src/main/java/com/example/internship/service/BookcaseService.java
@@ -1,9 +1,11 @@
 package com.example.internship.service;
 
 import com.example.internship.dao.Bookcase;
+import com.example.internship.dao.Shelf;
 import com.example.internship.repository.BookcaseRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
@@ -13,8 +15,23 @@ public class BookcaseService {
 
     private final BookcaseRepository bookcaseService;
 
-    public void delete(long id) {
-        bookcaseService.deleteById(id);
+    @Transactional
+    public void delete(Long id) {
+        Optional<Bookcase> optionalBookcase = bookcaseService.findById(id);
+
+        if (optionalBookcase.isPresent()) {
+            Bookcase bookcase = optionalBookcase.get();
+
+            for (Shelf shelf : bookcase.getShelfs()) {
+                shelf.setBookcase(null);
+            }
+            bookcase.getShelfs().clear();
+
+            bookcaseService.save(bookcase);
+            bookcaseService.deleteById(id);
+        } else {
+            throw new RuntimeException("Шкаф с ID " + id + " не найден");
+        }
     }
 
     public Bookcase save(Bookcase bookcase){
