+++ src/main/java/com/example/internship/service/ShelfService.java
@@ -1,9 +1,12 @@
 package com.example.internship.service;
 
+import com.example.internship.dao.Book;
+import com.example.internship.dao.Client;
 import com.example.internship.dao.Shelf;
 import com.example.internship.repository.ShelfRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
@@ -13,8 +16,29 @@ public class ShelfService {
 
     private final ShelfRepository shelfRepository;
 
-    public void delete(long id) {
-        shelfRepository.deleteById(id);
+    @Transactional
+    public void delete(Long id) {
+        Optional<Shelf> shelfOptional = shelfRepository.findById(id);
+
+        if (shelfOptional.isPresent()) {
+            Shelf shelf = shelfOptional.get();
+
+            for (Book book : shelf.getBooks()) {
+                book.setShelf(null);
+            }
+            shelf.getBooks().clear();
+
+            if (shelf.getBookcase() != null) {
+                shelf.getBookcase().getShelfs().remove(shelf);
+                shelf.setBookcase(null);
+            }
+
+            shelfRepository.save(shelf);
+
+            shelfRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Клиент с ID " + id + " не найден");
+        }
     }
 
     public Shelf save(Shelf shelf){
